BRS-760: Creating function for making a reservation object
